+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## HEAD (Unreleased)
 
+- Feature: autosize input width
+- Feature: fill input style
+- Feature: sm, md, lg input sizing
+
 ## 29.1.0 (2020-06-29)
 
 - Feature: Show step labels in horizontal stepper