+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -5,6 +5,8 @@
 - Feature: autosize input width
 - Feature: fill input style
 - Feature: sm, md, lg input sizing
+- Bug: make context drawer compatible with components
+- Bug: fix ngx-drawer-content scrolling css issue
 
 ## 29.1.0 (2020-06-29)