+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -6,7 +6,7 @@
 - Feature: fill input style
 - Feature: sm, md, lg input sizing
 - Bug: make context drawer compatible with components
-- Bug: fix ngx-drawer-content scrolling css issue
+- Bug: fix ngx-drawer-content scrolling scss issue
 
 ## 29.1.0 (2020-06-29)