+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## HEAD (Unreleased)
 
 - Feature: support `cssClass` in alert/config options
+- Feature: ngx-tip component
 - Feature: autosize input width
 - Feature: fill input style
 - Feature: sm, md, lg input sizing